Correct dispose of flutter switch widget
Hi, first of all, thanks for publishing this package, it is awesome. In a project that I use it, I was getting an exception when I tried to force the widget to rebuilt. The exception was exception: disposed with an active ticker. The solution is to dispose the animation controller, so there you go.